% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_angle.R
\name{add_angle}
\alias{add_angle}
\title{Adds a new column called angle. The angle is given by
$$ angle  = year di(# days in year)*360$$
This anlge is for plotting later.}
\usage{
add_angle(stn_data)
}
\arguments{
\item{stn_data}{A data frame of station data, must include the variable
date}
}
\value{
station data with an addition column called angle
}
\description{
Adds a new column called angle. The angle is given by
$$ angle  = year di(# days in year)*360$$
This anlge is for plotting later.
}
