% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scaled_rainfall.R
\name{add_scaled_rainfall}
\alias{add_scaled_rainfall}
\title{Adds a new column to the station data of the rainfall.
This column rescales the rainfall data relative to the
maximum rainfall on record. This rescale is for plotting later.}
\usage{
add_scaled_rainfall(stn_data, power = 2)
}
\arguments{
\item{stn_data}{A data frame of station data, must include the variable
prcp (precipitation)}

\item{power}{The power we use to rescale the rainfall (Default = 2)}
}
\value{
station data with an addition column of scaled rainfall
}
\description{
Adds a new column to the station data of the rainfall.
This column rescales the rainfall data relative to the
maximum rainfall on record. This rescale is for plotting later.
}
