% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_colour_palette.R
\name{create_colour_palette}
\alias{create_colour_palette}
\title{Function creates the colour palette for plotting the temperature bars}
\usage{
create_colour_palette(colors, pal_len = 100)
}
\arguments{
\item{colors}{a vector of colours for a colour scale ranging from low to high
(Default is colors <- c("darkturquoise", "deepskyblue", "deepskyblue1",
"deepskyblue1", "deepskyblue2", "deepskyblue2", "deepskyblue3", "maroon1",
"firebrick1") and was chosen to be consistent with the original reference
artwork)}

\item{pal_len}{The length of the temperature palette returned.
(Default is 100)}
}
\value{
A colour paletter in vector form
}
\description{
Function creates the colour palette for plotting the temperature bars
}
