% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temperature_bar_plot.R
\name{create_temperature_bar_plot}
\alias{create_temperature_bar_plot}
\title{Function creates a bar plot of temperature.
Instead of a normal bar plot, each bars is represented as line segments.
Each of these line segments are coloured differently to reflect the
reference climatology.}
\usage{
create_temperature_bar_plot(plot_data, mid_quantiles, tmin_min, tmax_max,
  color_palette, white_shift = 0.25)
}
\arguments{
\item{plot_data}{data frame of station data; must have columns angle, tmin and tmax.
Angle corresponds to the day of the year, see \code{add_angle()}. If rainfall circles
are to be added
then plot_data must also have a column called rf_scale, for rescaled
rainfall, see \code{add_scaled_rainfall()}.}

\item{mid_quantiles}{This is a vector, that gives the y-coordindates for the
start and end of each line segemnt.
To use segments relative to reference climatology use \code{get_climatology_quantiles()}.}

\item{tmin_min}{Double giving the coldest temperature on record.
Needed for plotting line segments outside the range of the reference climatology.}

\item{color_palette}{This will be used to plot the line segments of each quantile.
(Default is generated by create_color_palette()).}

\item{white_shift}{Optional double that shifts the bar plot by this amount
for aesthetic reasons (defult is 0.25).}

\item{tmax_min}{Double giving the hottest temperature on record.
Needed for plotting line segments outside the range of the reference climatology.}
}
\value{
Returns a bar plot where each bar corresponds to a day of the
year. The start of the bar is the minimum temperature for that day and
end of the bar is the maximum temperature. The bar is created using line
segements which are coloured according the the reference climatology, given
by mid_quantiles.
}
\description{
Function creates a bar plot of temperature.
Instead of a normal bar plot, each bars is represented as line segments.
Each of these line segments are coloured differently to reflect the
reference climatology.
}
\examples{
tmin = sin(seq(0, 2*pi, length.out = 360)) + rnorm(360, 20, 3)
plot_data = data.frame(angle = 1:360,
    tmin = tmin,
    tmax = tmin + 10)
mid_quantiles = seq(18, 28, length.out = 10)
tmin_min = min(plot_data$tmin) - 1
tmax_max = max(plot_data$tmax) + 1
create_temperature_bar_plot(plot_data, mid_quantiles,
   tmin_min, tmax_max)

}
