% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climatology_quantiles.R
\name{get_climatology_quantiles}
\alias{get_climatology_quantiles}
\title{Returns the quantiels from the period of reference climatology for temperature.
The quantiles are defined relative to the difference between TMIN and TMAX.}
\usage{
get_climatology_quantiles(stn_data, last_year, break_vals = seq(0, 1,
  0.025))
}
\arguments{
\item{stn_data}{A data frame of station data, must include variables;
date, tmax and tmin. Output of function \code{get_station_data()}.}

\item{last_year}{The numeric value for the year that ends of the
cliamte reference period. We will estimate the quantiles for the
reference climatology from this period.
If not provided the default value is the last year on the record.}

\item{break_vals}{A vector given the quantiles to estimate using the
reference climatology.
These values will correspond directly to a colour scale later.
(Default is seq(0,1,0.025))
These values control the smoothness of colour scale, but be warned too many
breaks will make the plot object needlessly complex.}
}
\value{
Returns empirical quantiles for \code{(tmax - tmin)/2 + tmin}
for the reference climatology period
}
\description{
Returns the quantiels from the period of reference climatology for temperature.
The quantiles are defined relative to the difference between TMIN and TMAX.
}
