% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year_data.R
\name{get_year_data}
\alias{get_year_data}
\title{Filters the station data by the given year}
\usage{
get_year_data(stn_data, year_val)
}
\arguments{
\item{stn_data}{A data frame of station data}

\item{year_val}{A numeric value for the year that will be used to filter
the data}
}
\value{
only the station data for the given year
}
\description{
Filters the station data by the given year
}
