% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_angle.R
\name{add_angle}
\alias{add_angle}
\title{Add a new column called angle}
\usage{
add_angle(stn_data)
}
\arguments{
\item{stn_data}{A data frame of station data, must include the variable
date}
}
\value{
returns the \code{stn_data} input with an addition column called angle
}
\description{
Adds a new column called angle to the input data frame. The angle is given by
(day in year)/(# days in year)*360.
This angle is used later in plotting functions.
}
\examples{
date_seq = seq(lubridate::as_date("2019-01-01"),
               lubridate::as_date("2019-01-31"),
               by = "days")
stn_data = data.frame(date = date_seq)
add_angle(stn_data)

}
